+++ Plugins/Events/Events.cpp
@@ -2,7 +2,7 @@
 #include "API/CExoString.hpp"
 #include "API/CVirtualMachine.hpp"
 #include "API/Globals.hpp"
-#include "API/Version.hpp"
+#include "API/Constants.hpp"
 #include "Events/AssociateEvents.hpp"
 #include "Events/BarterEvents.hpp"
 #include "Events/ClientEvents.hpp"
@@ -31,8 +31,11 @@
 #include "ViewPtr.hpp"
 #include <algorithm>
 #include <regex>
+#include <string>
 
 using namespace NWNXLib;
+using namespace NWNXLib::API;
+using namespace NWNXLib::API::Constants;
 
 static ViewPtr<Events::Events> g_plugin;
 
@@ -63,6 +66,7 @@ Events::Events(const Plugin::CreateParams& params)
     if (g_plugin == nullptr) // :(
         g_plugin = this;
 
+    // TODO-64bit: Bring this stuff more in line with how the other plugins handle this
     GetServices()->m_events->RegisterEvent("SUBSCRIBE_EVENT", std::bind(&Events::OnSubscribeEvent, this, std::placeholders::_1));
     GetServices()->m_events->RegisterEvent("PUSH_EVENT_DATA", std::bind(&Events::OnPushEventData, this, std::placeholders::_1));
     GetServices()->m_events->RegisterEvent("SIGNAL_EVENT", std::bind(&Events::OnSignalEvent, this, std::placeholders::_1));
@@ -70,6 +74,9 @@ Events::Events(const Plugin::CreateParams& params)
     GetServices()->m_events->RegisterEvent("SKIP_EVENT", std::bind(&Events::OnSkipEvent, this, std::placeholders::_1));
     GetServices()->m_events->RegisterEvent("EVENT_RESULT", std::bind(&Events::OnEventResult, this, std::placeholders::_1));
     GetServices()->m_events->RegisterEvent("GET_CURRENT_EVENT", std::bind(&Events::OnGetCurrentEvent, this, std::placeholders::_1));
+    GetServices()->m_events->RegisterEvent("ToggleDispatchListMode", std::bind(&Events::ToggleDispatchListMode, this, std::placeholders::_1));
+    GetServices()->m_events->RegisterEvent("AddObjectToDispatchList", std::bind(&Events::AddObjectToDispatchList, this, std::placeholders::_1));
+    GetServices()->m_events->RegisterEvent("RemoveObjectFromDispatchList", std::bind(&Events::RemoveObjectFromDispatchList, this, std::placeholders::_1));
 
     GetServices()->m_messaging->SubscribeMessage("NWNX_EVENT_SIGNAL_EVENT",
         [](const std::vector<std::string> message)
@@ -121,7 +128,6 @@ void Events::PushEventData(const std::string tag, const std::string data)
     g_plugin->m_eventData.top().m_EventDataMap[tag] = std::move(data);
 }
 
-
 std::string Events::GetEventData(const std::string tag)
 {
     std::string retVal;
@@ -145,7 +151,7 @@ std::string Events::GetEventData(const std::string tag)
     return retVal;
 }
 
-bool Events::SignalEvent(const std::string& eventName, const API::Types::ObjectID target, std::string *result)
+bool Events::SignalEvent(const std::string& eventName, const Types::ObjectID target, std::string *result)
 {
     bool skipped = false;
 
@@ -155,20 +161,35 @@ bool Events::SignalEvent(const std::string& eventName, const API::Types::ObjectI
 
     for (const auto& script : g_plugin->m_eventMap[eventName])
     {
-        LOG_DEBUG("Dispatching notification for event '%s' to script '%s'.", eventName.c_str(), script.c_str());
-        API::CExoString scriptExoStr = script.c_str();
+        auto DispatchEvent = [&]() -> void {
+            LOG_DEBUG("Dispatching notification for event '%s' to script '%s'.", eventName.c_str(), script.c_str());
+            API::CExoString scriptExoStr = script.c_str();
 
-        ++g_plugin->m_eventDepth;
-        API::Globals::VirtualMachine()->RunScript(&scriptExoStr, target, 1);
+            ++g_plugin->m_eventDepth;
+            API::Globals::VirtualMachine()->RunScript(&scriptExoStr, target, 1);
 
-        skipped |= g_plugin->m_eventData.top().m_Skipped;
+            skipped |= g_plugin->m_eventData.top().m_Skipped;
 
-        if (result)
+            if (result)
+            {
+                *result = g_plugin->m_eventData.top().m_Result;
+            }
+
+            --g_plugin->m_eventDepth;
+        };
+
+        auto eventDispatchList = g_plugin->m_dispatchList.find(eventName + script);
+        if (eventDispatchList != g_plugin->m_dispatchList.end())
         {
-            *result = g_plugin->m_eventData.top().m_Result;
+            if(eventDispatchList->second.find(target) != eventDispatchList->second.end())
+            {
+                DispatchEvent();
+            }
         }
-
-        --g_plugin->m_eventDepth;
+        else
+        {
+            DispatchEvent();
+        }
     }
 
     g_plugin->m_eventData.pop();
@@ -204,7 +225,7 @@ void Events::RunEventInit(const std::string& eventName)
 
 }
 
-Services::Events::ArgumentStack Events::OnSubscribeEvent(Services::Events::ArgumentStack&& args)
+ArgumentStack Events::OnSubscribeEvent(ArgumentStack&& args)
 {
     const auto event = Services::Events::ExtractArgument<std::string>(args);
     auto script = Services::Events::ExtractArgument<std::string>(args);
@@ -220,36 +241,36 @@ Services::Events::ArgumentStack Events::OnSubscribeEvent(Services::Events::Argum
     LOG_INFO("Script '%s' subscribed to event '%s'.", script.c_str(), event.c_str());
     eventVector.emplace_back(std::move(script));
 
-    return Services::Events::ArgumentStack();
+    return ArgumentStack();
 }
 
-Services::Events::ArgumentStack Events::OnPushEventData(Services::Events::ArgumentStack&& args)
+ArgumentStack Events::OnPushEventData(ArgumentStack&& args)
 {
     const auto tag = Services::Events::ExtractArgument<std::string>(args);
     const auto data = Services::Events::ExtractArgument<std::string>(args);
     PushEventData(tag, data);
-    return Services::Events::ArgumentStack();
+    return ArgumentStack();
 }
 
-Services::Events::ArgumentStack Events::OnSignalEvent(Services::Events::ArgumentStack&& args)
+ArgumentStack Events::OnSignalEvent(ArgumentStack&& args)
 {
     const auto event = Services::Events::ExtractArgument<std::string>(args);
-    const auto object = Services::Events::ExtractArgument<API::Types::ObjectID>(args);
+    const auto object = Services::Events::ExtractArgument<Types::ObjectID>(args);
     bool signalled = SignalEvent(event, object);
-    Services::Events::ArgumentStack stack;
+    ArgumentStack stack;
     Services::Events::InsertArgument(stack, signalled ? 1 : 0);
     return stack;
 }
 
-Services::Events::ArgumentStack Events::OnGetEventData(Services::Events::ArgumentStack&& args)
+ArgumentStack Events::OnGetEventData(ArgumentStack&& args)
 {
     std::string data = GetEventData(Services::Events::ExtractArgument<std::string>(args));
-    Services::Events::ArgumentStack stack;
+    ArgumentStack stack;
     Services::Events::InsertArgument(stack, data);
     return stack;
 }
 
-Services::Events::ArgumentStack Events::OnSkipEvent(Services::Events::ArgumentStack&&)
+ArgumentStack Events::OnSkipEvent(ArgumentStack&&)
 {
     if (m_eventDepth == 0 || m_eventData.empty())
     {
@@ -259,10 +280,10 @@ Services::Events::ArgumentStack Events::OnSkipEvent(Services::Events::ArgumentSt
 
     LOG_DEBUG("Skipping last event.");
 
-    return Services::Events::ArgumentStack();
+    return ArgumentStack();
 }
 
-Services::Events::ArgumentStack Events::OnEventResult(Services::Events::ArgumentStack&& args)
+ArgumentStack Events::OnEventResult(ArgumentStack&& args)
 {
     if (m_eventDepth == 0 || m_eventData.empty())
     {
@@ -274,10 +295,10 @@ Services::Events::ArgumentStack Events::OnEventResult(Services::Events::Argument
 
     LOG_DEBUG("Received event result '%s'.", data.c_str());
 
-    return Services::Events::ArgumentStack();
+    return ArgumentStack();
 }
 
-Services::Events::ArgumentStack Events::OnGetCurrentEvent(Services::Events::ArgumentStack&&)
+ArgumentStack Events::OnGetCurrentEvent(ArgumentStack&&)
 {
     std::string retVal;
 
@@ -290,8 +311,66 @@ Services::Events::ArgumentStack Events::OnGetCurrentEvent(Services::Events::Argu
         retVal = g_plugin->m_eventData.top().m_EventName;
     }
 
-    Services::Events::ArgumentStack stack;
+    ArgumentStack stack;
     Services::Events::InsertArgument(stack, retVal);
+    return stack;
+}
+
+ArgumentStack Events::ToggleDispatchListMode(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+
+    const auto eventName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!eventName.empty());
+    const auto scriptName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!scriptName.empty());
+    const bool bEnable = Services::Events::ExtractArgument<int32_t>(args) != 0;
+
+    if (bEnable)
+        g_plugin->m_dispatchList[eventName+scriptName];
+    else
+        g_plugin->m_dispatchList.erase(eventName+scriptName);
+
+    return stack;
+}
+
+ArgumentStack Events::AddObjectToDispatchList(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+
+    const auto eventName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!eventName.empty());
+    const auto scriptName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!scriptName.empty());
+    const auto oidObject = Services::Events::ExtractArgument<Types::ObjectID>(args);
+      ASSERT_OR_THROW(oidObject != Constants::OBJECT_INVALID);
+
+    auto eventDispatchList = g_plugin->m_dispatchList.find(eventName+scriptName);
+    if (eventDispatchList != g_plugin->m_dispatchList.end())
+    {
+        eventDispatchList->second.insert(oidObject);
+    }
+
+    return stack;
+}
+
+ArgumentStack Events::RemoveObjectFromDispatchList(ArgumentStack&& args)
+{
+    ArgumentStack stack;
+
+    const auto eventName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!eventName.empty());
+    const auto scriptName = Services::Events::ExtractArgument<std::string>(args);
+      ASSERT_OR_THROW(!scriptName.empty());
+    const auto oidObject = Services::Events::ExtractArgument<Types::ObjectID>(args);
+      ASSERT_OR_THROW(oidObject != Constants::OBJECT_INVALID);
+
+    auto eventDispatchList = g_plugin->m_dispatchList.find(eventName+scriptName);
+    if (eventDispatchList != g_plugin->m_dispatchList.end())
+    {
+        eventDispatchList->second.erase(oidObject);
+    }
+
     return stack;
 }